ban gop hoan chinh

+++ database/migrations/2025_03_08_155953_create_variant_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('idproduct')->constrained('products')->onDelete('cascade');
             $table->string('code',255);
-            $table->double('price',8,2)->nullable();
+            $table->decimal('price', 10, 2)->nullable();
             $table->integer('quantity');
             $table->timestamps();
         });
